Failide üleslaadimine

+++ docker_start.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 
-cp -R -u -p /src/config_dist.js /config/config.js
+if [ ! -f /config/config.js ]; then
+    cp -R -u -p /src/config_dist.js /config/config.js
+fi
+
 ln -s /config/config.js /src/config.js
 forever start /src/app.js
 forever list
